+++ feature/center/build.gradle.kts
@@ -27,10 +27,10 @@ kotlin{
             implementation(compose.components.uiToolingPreview)
             implementation(compose.ui)
             implementation(libs.androidx.paging.common)
+            implementation(libs.androidx.paging.compose)
             implementation(libs.coil.kt.compose)
         }
         androidMain.dependencies{
-            implementation(libs.androidx.paging.compose)
             implementation(compose.ui)
             implementation(compose.material3)
             implementation(libs.androidx.compose.foundation)
