+++ feature/center/src/commonMain/kotlin/com/mifos/feature/center/centerList/ui/CenterListScreen.kt
@@ -10,8 +10,10 @@
 package com.mifos.feature.center.centerList.ui
 
 import androidclient.feature.center.generated.resources.Res
+import androidclient.feature.center.generated.resources.feature_center_error_loading_centers
 import androidclient.feature.center.generated.resources.feature_center_failed_to_load_db_centers
 import androidclient.feature.center.generated.resources.feature_center_ic_done_all_black_24dp
+import androidclient.feature.center.generated.resources.feature_center_no_more_centers
 import androidclient.feature.center.generated.resources.feature_center_sync
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.ExperimentalFoundationApi
@@ -20,6 +22,8 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
@@ -31,6 +35,7 @@ import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.ListItemDefaults
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedCard
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
@@ -49,13 +54,18 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.paging.LoadState
 import androidx.paging.PagingData
+import androidx.paging.compose.collectAsLazyPagingItems
 import coil3.compose.AsyncImage
 import com.mifos.core.designsystem.component.MifosScaffold
 import com.mifos.core.designsystem.component.MifosSweetError
 import com.mifos.core.designsystem.icon.MifosIcons
 import com.mifos.core.designsystem.theme.DesignToken
+import com.mifos.core.ui.components.MifosPagingAppendProgress
 import com.mifos.core.ui.components.MifosProgressIndicator
 import com.mifos.core.ui.components.SelectionModeTopAppBar
 import com.mifos.feature.center.syncCentersDialog.SyncCenterDialogScreen
@@ -182,6 +192,7 @@ internal fun CenterListScreen(
                             onCenterSelect = {
                                 onCenterSelect(it)
                             },
+                            modifier = Modifier,
                         )
                     }
 
@@ -234,14 +245,87 @@ class SelectedItemsState(initialSelectedItems: List<CenterEntity> = emptyList())
 }
 
 @Composable
-expect fun CenterListContent(
+fun CenterListContent(
     state: CenterListUiState,
     isInSelectionMode: Boolean,
     selectedItems: SelectedItemsState,
     onRefresh: () -> Unit,
     onCenterSelect: (Int) -> Unit,
-    modifier: Modifier = Modifier,
-)
+    modifier: Modifier,
+) {
+    if (state is CenterListUiState.CenterList) {
+        val centerPagingList = state.centers.collectAsLazyPagingItems()
+        when (centerPagingList.loadState.refresh) {
+            is LoadState.Error -> {
+                MifosSweetError(message = stringResource(Res.string.feature_center_error_loading_centers)) {
+                    onRefresh()
+                }
+            }
+
+            is LoadState.Loading -> MifosProgressIndicator()
+
+            is LoadState.NotLoading -> Unit
+        }
+
+        LazyColumn(
+            modifier = modifier.fillMaxSize().padding(horizontal = 16.dp),
+            verticalArrangement = Arrangement.spacedBy(12.dp),
+        ) {
+            items(
+                count = centerPagingList.itemCount,
+                key = {
+                    centerPagingList[it]?.id ?: it
+                },
+            ) { index ->
+                val center = centerPagingList[index]!!
+
+                CenterCard(
+                    center = center,
+                    selected = selectedItems.contains(center),
+                    isInSelectionMode = selectedItems.size() > 0,
+                    onSelect = {
+                        if (selectedItems.contains(it)) {
+                            selectedItems.remove(it)
+                        } else {
+                            selectedItems.add(it)
+                        }
+                    },
+                    onClick = {
+                        onCenterSelect(it.id ?: 0)
+                    },
+                )
+            }
+
+            when (centerPagingList.loadState.append) {
+                is LoadState.Error -> {}
+
+                is LoadState.Loading -> {
+                    item {
+                        MifosPagingAppendProgress()
+                    }
+                }
+
+                is LoadState.NotLoading -> Unit
+            }
+            when (centerPagingList.loadState.append.endOfPaginationReached) {
+                true -> {
+                    item {
+                        Text(
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .padding(6.dp),
+                            text = stringResource(Res.string.feature_center_no_more_centers),
+                            style = MaterialTheme.typography.bodyMedium,
+                            textAlign = TextAlign.Center,
+                        )
+                    }
+                }
+
+                false -> Unit
+            }
+        }
+    }
+}
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
